+++ Example/Emission/AppSetup.h
@@ -17,6 +17,7 @@
 
 @property (copy, nonatomic, readonly) NSString *gravityURL;
 @property (copy, nonatomic, readonly) NSString *metaphysicsURL;
+@property (copy, nonatomic, readonly) NSString *predictionURL;
 @property (copy, nonatomic, readonly) NSString *packagerURL;
 
 @property (nonatomic, strong, readonly) NSString *emissionLoadedFromString;